requirements.txt
colorama==0.4.6
faiss-cpu==1.9.0
Flask==3.1.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
itsdangerous==2.2.0
Jinja2==3.1.4
langchain-core==0.3.19
langchain-google-genai==2.0.4
MarkupSafe==3.0.2
numpy==2.1.3
orjson==3.10.11
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
